# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /my/sites-summary:
    get:
      operationId: getMySitesSummary
      tags: [my]
      summary: Plan/fact summary per assigned site for the current user
      parameters:
        - name: from
          in: query
          required: false
          description: Date YYYY-MM-DD
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          required: false
          description: Date YYYY-MM-DD
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Per-site plan/fact rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MySiteSummary"
...
components:
  schemas:
...
    MySiteSummary:
      type: object
      properties:
        siteId:
          type: string
        siteName:
          type: string
        address:
          type: string
        driverUserId:
          type: [string, "null"]
          format: uuid
        driver:
          type: string
        planned:
          type: integer
          description: Всего плановых доставок за период
        done:
          type: integer
          description: Выполнено (есть actualDate)
        overdue:
          type: integer
          description: Просрочено (дата прошла, факта нет)
      required:
        [
          siteId,
          siteName,
          address,
          driverUserId,
          driver,
          planned,
          done,
          overdue,
        ]
...